--- routing_server.js
@@ -1,15 +1,0 @@
-const express = require("express");
-const app = express();
-const port = 3000;
-
-app.get("/", function (req, res) {
-  res.send("Hello World!");
-});
-
-app.get("/unique", function (req, res) {
-  res.send("This is a different page!");
-});
-
-app.listen(port, function () {
-  console.log(`Example app listening on port ${port}!`);
-});
