--- views/pages/enter_book_info.ejs
@@ -1,24 +1,0 @@
-<!DOCTYPE html> 
-<html> 
-<head> 
-	<title>Add New Book</title> 
-
-	<style type="text/css" media="screen"> 
-		body { 
-			background-color: skyblue; 
-			text-decoration-color: white; 
-		} 
-	</style> 
-</head> 
-
-<body> 
-	<h1>Enter a new book to add.</h1>
-	<form action="./add_book" method="post">
-		<label for="title">Book title:</label>
-		<input type="text" id="title" name="title"><br>
-		<label for="author">Author:</label>
-		<input type="text" id="author" name="author"><br>
-		<input type="submit" value="Submit">
-	</form> 
-</body> 
-</html>
