+++ views/pages/home.ejs
@@ -1,17 +1,13 @@
-<!DOCTYPE html> 
-<html> 
-<head> 
-	<title>Home Page</title> 
+<!DOCTYPE html>
+<html>
+  <head>
+    <title><%= title %></title>
+    <link rel='stylesheet' href='/stylesheets/style.css' />
+  </head>
+  <body>
+    <h1><%= title %></h1>
+    <p>Welcome to my EJS <%= title %></p>
 
-	<style type="text/css" media="screen"> 
-		body { 
-			background-color: skyblue; 
-			text-decoration-color: white; 
-		} 
-	</style> 
-</head> 
-
-<body> 
-	<center><h1>This is a static home page.</h1></center> 
-</body> 
+	<img src="/images/cow.jpg">
+  </body>
 </html>
