--- views/pages/home_dynamic.ejs
@@ -1,30 +1,0 @@
-<!DOCTYPE html> 
-<html> 
-<head> 
-	<title>Home Page</title> 
-
-	<style type="text/css" media="screen"> 
-		body { 
-			background-color: skyblue; 
-			text-decoration-color: white; 
-		}
-		h1 { 
-			font-size: 7em; 
-		}
-		p { 
-			font-size: 4em; 
-		}
-	</style> 
-</head> 
-
-<body> 
-	<center><h1>This is a dynamic home page.</h1>
-	<p>Hello <%= user.name %>!</br>
-	<small>I see you like <font color="<%= user.favorite_color %>"><%= user.favorite_color %></font></small>
-	</p>
-	<% if (user.isGreatestTeacher) { %>
-	<p style="font-size: 1em;">I also see you are the greatest teacher of all time. I bow in your presence.</p>
-	<% } %>
-	</center> 
-</body> 
-</html>
