--- views/pages/home_lists.ejs
@@ -1,31 +1,0 @@
-<!DOCTYPE html> 
-<html> 
-<head> 
-	<title>Home Page</title> 
-
-	<style type="text/css" media="screen"> 
-		body { 
-			background-color: skyblue; 
-			text-decoration-color: white; 
-		}
-	</style> 
-</head> 
-
-<body> 
-	<center><h1>This page demonstrates lists.</h1>
-	<ul>
-        <% books.forEach((book)=> { %>
-            <li>
-                <h2>
-                    <%= book.title %>
-                </h2>
-                <p>
-                    <%= book.author %>
-                </p>
-            </li>
-            <hr />
-        <% }) %>
-    </ul>
-	</center> 
-</body> 
-</html>
